tools: limit MacOS notarization with gon to Xcode version < 13.0

+++ tools/osx-notarize.sh
@@ -1,34 +1,44 @@
 #!/bin/sh
-
-# Uses gon, from https://github.com/mitchellh/gon, to notarize a generated node-<version>.pkg file
-# with Apple for installation on macOS Catalina and later as validated by Gatekeeper.
+# Notarize a generated node-<version>.pkg file is an Apple requirement for installation on macOS Catalina and later, as validated by Gatekeeper.
+# Uses gon, from https://github.com/mitchellh/gon (Xcode version < 13.0.)
+# Uses notarytool (Xcode >= 13.0).
 
 set -e
 
-gon_version="0.2.2"
-gon_exe="${HOME}/.gon/gon_${gon_version}"
+xcode_version=$(xcodebuild -version | awk '/Xcode/ {print $2}')
 
-pkgid="$1"
+if [[ "$xcode_version" < "13.0" ]]; then
+  echo "Notarization process is done with gon."
 
-[ -z "$pkgid" ] && \
-  echo "Usage: $0 <pkgid>" \
-  exit 1
+  gon_version="0.2.2"
+  gon_exe="${HOME}/.gon/gon_${gon_version}"
 
-# shellcheck disable=SC2154
-[ -z "$NOTARIZATION_ID" ] && \
-  echo "No NOTARIZATION_ID environment var. Skipping notarization." \
-  exit 0
+  pkgid="$1"
 
-set -x
+  [ -z "$pkgid" ] && \
+    echo "Usage: $0 <pkgid>" \
+    exit 1
 
-mkdir -p "${HOME}/.gon/"
+  # shellcheck disable=SC2154
+  [ -z "$NOTARIZATION_ID" ] && \
+    echo "No NOTARIZATION_ID environment var. Skipping notarization." \
+    exit 0
 
-if [ ! -f "${gon_exe}" ]; then
-  curl -sL "https://github.com/mitchellh/gon/releases/download/v${gon_version}/gon_${gon_version}_macos.zip" -o "${gon_exe}.zip"
-  (cd "${HOME}/.gon/" && rm -f gon && unzip "${gon_exe}.zip" && mv gon "${gon_exe}")
+  set -x
+
+  mkdir -p "${HOME}/.gon/"
+
+  if [ ! -f "${gon_exe}" ]; then
+    curl -sL "https://github.com/mitchellh/gon/releases/download/v${gon_version}/gon_${gon_version}_macos.zip" -o "${gon_exe}.zip"
+    (cd "${HOME}/.gon/" && rm -f gon && unzip "${gon_exe}.zip" && mv gon "${gon_exe}")
+  fi
+
+  sed -e "s/{{appleid}}/${NOTARIZATION_ID}/" -e "s/{{pkgid}}/${pkgid}/" tools/osx-gon-config.json.tmpl \
+    > gon-config.json
+
+  "${gon_exe}" -log-level=info gon-config.json
+
+else
+  echo "Notarization process is done with Notarytool."
+  # @TODO: Implement notarization with notarytool.
 fi
-
-sed -e "s/{{appleid}}/${NOTARIZATION_ID}/" -e "s/{{pkgid}}/${pkgid}/" tools/osx-gon-config.json.tmpl \
-  > gon-config.json
-
-"${gon_exe}" -log-level=info gon-config.json
